moved metrics to top level [ci skip]
