+++ openpnm/metrics/GenericMetric.py
