+++ openpnm/metrics/RelativePermeability.py
